requirements.txt

certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
idna==2.8
numpy==2.2.3
pandas==2.2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
tqdm==4.62.3
twilio==7.14.0
tzdata==2025.1
urllib3==2.3.0
